+++ saleor/order/actions.py
@@ -11,6 +11,7 @@ from django.contrib.sites.models import Site
 from django.db import transaction
 
 from ..account.models import User
+from ..app.models import App
 from ..core.exceptions import AllocationError, InsufficientStock, InsufficientStockData
 from ..core.tracing import traced_atomic_transaction
 from ..core.transactions import transaction_with_commit_on_errors
@@ -379,7 +380,7 @@ def handle_fully_paid_order(
     if order_info.customer_email:
         send_payment_confirmation(order_info, manager)
         if utils.order_needs_automatic_fulfillment(order_info.lines_data):
-            automatically_fulfill_digital_lines(order_info, manager)
+            automatically_fulfill_digital_lines(order_info, manager, user, app)
 
     if site_settings is None:
         site_settings = Site.objects.get_current().settings
@@ -537,6 +538,8 @@ def order_fulfilled(
     gift_card_lines_info: list[GiftCardLineData],
     site_settings: "SiteSettings",
     notify_customer=True,
+    auto=False,
+    manually_approved=False,
     webhook_event_map: dict[str, set["Webhook"]] | None = None,
 ):
     from ..giftcard.utils import gift_cards_create
@@ -550,25 +553,33 @@ def order_fulfilled(
     # events are successfully created
     with traced_atomic_transaction():
         update_order_status(order)
-        gift_cards_create(
-            order,
-            gift_card_lines_info,
-            site_settings,
-            user,
-            app,
-            manager,
-        )
+        if gift_card_lines_info:
+            gift_cards_create(
+                order,
+                gift_card_lines_info,
+                site_settings,
+                user,
+                app,
+                manager,
+            )
         events.fulfillment_fulfilled_items_event(
-            order=order, user=user, app=app, fulfillment_lines=fulfillment_lines
+            order=order,
+            user=user,
+            app=app,
+            fulfillment_lines=fulfillment_lines,
+            auto=auto,
         )
         webhook_events = [WebhookEventAsyncType.ORDER_UPDATED]
         for fulfillment in fulfillments:
             call_event(manager.fulfillment_created, fulfillment, notify_customer)
 
-        if order.status == OrderStatus.FULFILLED:
+        order_fulfilled = order.status == OrderStatus.FULFILLED
+        if order_fulfilled:
             webhook_events.append(WebhookEventAsyncType.ORDER_FULFILLED)
+
+        if order_fulfilled or manually_approved:
             for fulfillment in fulfillments:
-                call_event(manager.fulfillment_approved, fulfillment)
+                call_event(manager.fulfillment_approved, fulfillment, notify_customer)
 
         call_order_events(
             manager,
@@ -578,8 +589,13 @@ def order_fulfilled(
         )
     if notify_customer:
         for fulfillment in fulfillments:
-            send_fulfillment_confirmation_to_customer(
-                order, fulfillment, user, app, manager
+            call_event(
+                send_fulfillment_confirmation_to_customer,
+                order,
+                fulfillment,
+                user,
+                app,
+                manager,
             )
 
 
@@ -835,28 +851,17 @@ def approve_fulfillment(
     notify_customer=True,
     allow_stock_to_be_exceeded: bool = False,
 ):
-    from ..giftcard.utils import gift_cards_create
-
     with traced_atomic_transaction():
         fulfillment.status = FulfillmentStatus.FULFILLED
         fulfillment.save()
         order = fulfillment.order
-        if notify_customer:
-            send_fulfillment_confirmation_to_customer(
-                fulfillment.order, fulfillment, user, app, manager
-            )
-        events.fulfillment_fulfilled_items_event(
-            order=order,
-            user=user,
-            app=app,
-            fulfillment_lines=list(fulfillment.lines.all()),
-        )
         lines_to_fulfill = []
         gift_card_lines_info = []
         insufficient_stocks = []
-        for fulfillment_line in fulfillment.lines.all().prefetch_related(
+        fulfillment_lines = fulfillment.lines.all().prefetch_related(
             "order_line__variant"
-        ):
+        )
+        for fulfillment_line in fulfillment_lines:
             order_line = fulfillment_line.order_line
             variant = fulfillment_line.order_line.variant
 
@@ -898,32 +903,17 @@ def approve_fulfillment(
 
         _decrease_stocks(lines_to_fulfill, manager, allow_stock_to_be_exceeded)
         order.refresh_from_db()
-        update_order_status(order)
-
-        webhook_event_map = get_webhooks_for_multiple_events(
-            WEBHOOK_EVENTS_FOR_ORDER_FULFILLED
-        )
-        webhook_events = [WebhookEventAsyncType.ORDER_UPDATED]
-        call_event(manager.fulfillment_approved, fulfillment, notify_customer)
-        if order.status == OrderStatus.FULFILLED:
-            webhook_events.append(WebhookEventAsyncType.ORDER_FULFILLED)
-
-        call_order_events(
+        order_fulfilled(
+            [fulfillment],
+            user,
+            app,
+            list(fulfillment_lines),
             manager,
-            webhook_events,
-            order,
-            webhook_event_map=webhook_event_map,
+            gift_card_lines_info,
+            settings,
+            notify_customer,
+            manually_approved=True,
         )
-
-        if gift_card_lines_info:
-            gift_cards_create(
-                order,
-                gift_card_lines_info,
-                settings,
-                user,
-                app,
-                manager,
-            )
 
     return fulfillment
 
@@ -1077,7 +1067,10 @@ def fulfill_order_lines(
 
 
 def automatically_fulfill_digital_lines(
-    order_info: "OrderInfo", manager: "PluginsManager"
+    order_info: "OrderInfo",
+    manager: "PluginsManager",
+    user: User | None = None,
+    app: App | None = None,
 ):
     """Fulfill all digital lines which have enabled automatic fulfillment setting.
 
@@ -1125,6 +1118,9 @@ def automatically_fulfill_digital_lines(
 
         FulfillmentLine.objects.bulk_create(fulfillments)
         fulfill_order_lines(lines_info, manager)
+        events.fulfillment_fulfilled_items_event(
+            order=order, user=user, app=app, fulfillment_lines=fulfillments, auto=True
+        )
 
         send_fulfillment_confirmation_to_customer(
             order, fulfillment, user=order.user, app=None, manager=manager
@@ -1258,8 +1254,10 @@ def create_fulfillments(
     fulfillment_lines_for_warehouses: dict[UUID, list[OrderFulfillmentLineInfo]],
     manager: "PluginsManager",
     site_settings: "SiteSettings",
+    *,
     notify_customer: bool = True,
-    approved: bool = True,
+    auto: bool = False,
+    auto_approved: bool = True,
     allow_stock_to_be_exceeded: bool = False,
     tracking_number: str = "",
 ) -> list[Fulfillment]:
@@ -1287,7 +1285,8 @@ def create_fulfillments(
         notify_customer (bool): If `True` system send email about
             fulfillments to customer.
         site_settings (SiteSettings): Site settings used for creating gift cards.
-        approved (Boolean): fulfillments will have status fulfilled if it's True,
+        auto (Boolean): define if the fulfillment is automatic
+        auto_approved (Boolean): fulfillments will have status fulfilled if it's True,
             otherwise waiting_for_approval.
         allow_stock_to_be_exceeded (bool): If `True` then stock quantity could exceed.
             Default value is set to `False`.
@@ -1307,7 +1306,7 @@ def create_fulfillments(
     gift_card_lines_info: list[GiftCardLineData] = []
     status = (
         FulfillmentStatus.FULFILLED
-        if approved
+        if auto_approved
         else FulfillmentStatus.WAITING_FOR_APPROVAL
     )
 
@@ -1340,7 +1339,7 @@ def create_fulfillments(
                     order.channel.slug,
                     gift_card_lines_info,
                     manager,
-                    decrease_stock=approved,
+                    decrease_stock=auto_approved,
                     allow_stock_to_be_exceeded=allow_stock_to_be_exceeded,
                 )
             )
@@ -1349,28 +1348,25 @@ def create_fulfillments(
 
         FulfillmentLine.objects.bulk_create(fulfillment_lines)
         order.refresh_from_db()
-        if approved:
-            transaction.on_commit(
-                lambda: order_fulfilled(
-                    fulfillments,
-                    user,
-                    app,
-                    fulfillment_lines,
-                    manager,
-                    gift_card_lines_info,
-                    site_settings,
-                    notify_customer,
-                )
+        if auto_approved:
+            order_fulfilled(
+                fulfillments,
+                user,
+                app,
+                fulfillment_lines,
+                manager,
+                gift_card_lines_info,
+                site_settings,
+                notify_customer,
+                auto,
             )
         else:
-            transaction.on_commit(
-                lambda: order_awaits_fulfillment_approval(
-                    fulfillments,
-                    user,
-                    app,
-                    fulfillment_lines,
-                    manager,
-                )
+            order_awaits_fulfillment_approval(
+                fulfillments,
+                user,
+                app,
+                fulfillment_lines,
+                manager,
             )
     return fulfillments