+++ saleor/order/tests/test_order_actions.py
@@ -92,9 +92,12 @@ def test_handle_fully_paid_order_digital_lines(
 
     # then
     fulfillment = order.fulfillments.first()
-    event_order_paid = order.events.get()
-
-    assert event_order_paid.type == OrderEvents.ORDER_FULLY_PAID
+    assert set(order.events.values_list("type", flat=True)) == {
+        OrderEvents.ORDER_FULLY_PAID,
+        OrderEvents.FULFILLMENT_FULFILLED_ITEMS,
+    }
+    fulfilled_event = order.events.get(type=OrderEvents.FULFILLMENT_FULFILLED_ITEMS)
+    assert fulfilled_event.parameters["auto"] is True
 
     mock_send_payment_confirmation.assert_called_once_with(order_info, manager)
     send_fulfillment_confirmation_to_customer.assert_called_once_with(
@@ -1573,6 +1576,14 @@ def test_fulfill_digital_lines(
     assert line.digital_content_url
     assert mock_email_fulfillment.called
     mock_fulfillment_created.assert_called_once_with(fulfillment)
+
+    event = order_with_lines.events.filter(
+        type=OrderEvents.FULFILLMENT_FULFILLED_ITEMS
+    ).first()
+    assert event
+    assert set(event.parameters["fulfilled_items"]) == {
+        line.pk for line in fulfillment_lines
+    }
 
 
 @patch("saleor.order.actions.send_fulfillment_confirmation_to_customer")